requirements.txt
zmq
simplejson
kafka
msgpack-python
fluent-logger
redis
